# support leveldb.getmany()