… deps/leveldb/include/leveldb/c.h
… leveldb_test.go
… multikey_error.go