This adds the `GetMany()` API to our forked version of leveldb directly. This should be faster than calling `Get()` in go in a loop as each such call crossing go-to-c-and-back can be expensive in cgo. 

According to the benchmarks below, and if I am interpreting them correctly, `GetMany()` is about *20% faster* cpu-wise, and *25% higher* in i/o throughput than `Get()`, using fixed-key length of 20 bytes and fixed-value length of 128 bytes.

```
name             old time/op    new time/op    delta
DBGets/DBGets-2    14.2ms ± 8%    11.4ms ± 6%  -20.16%  (p=0.000 n=10+9)
name             old speed      new speed      delta
DBGets/DBGets-2  14.1MB/s ± 8%  17.6MB/s ± 6%  +25.08%  (p=0.000 n=10+9)
```